Move SearchResponseSchema to collections.schemas to fix build error
